March 31, 2024 Bot API 7.2
Integration with Business Accounts

Added the class BusinessConnection and updates about the connection or disconnection of the bot to a business account, represented by the field business_connection in the class Update.
Added updates about new messages in a business account connected to the bot, represented by the field business_message in the class Update.
Added updates about message edits in a business account connected to the bot, represented by the field edited_business_message in the class Update.
Added updates about message deletion in a business account connected to the bot, represented by the class BusinessMessagesDeleted and the field deleted_business_messages in the class Update.
Added the method getBusinessConnection.
Working on Behalf of Business Accounts

Added the parameter business_connection_id to the methods sendMessage, sendPhoto, sendVideo, sendAnimation, sendAudio, sendDocument, sendSticker, sendVideoNote, sendVoice, sendLocation, sendVenue, sendContact, sendPoll, sendDice, sendGame, and sendMediaGroup.
Added the parameter business_connection_id to the method sendChatAction.
Added the field business_connection_id to the class Message.
Added the field sender_business_bot to the class Message.
Information about Business Accounts

Added the class BusinessIntro and the field business_intro to the class Chat.
Added the class BusinessLocation and the field business_location to the class Chat.
Added the classes BusinessOpeningHours and BusinessOpeningHoursInterval and the field business_opening_hours to the class Chat.